+++ src/main/java/de/jeff_media/angelchest/Watchdog.java
@@ -39,7 +39,6 @@ public final class Watchdog {
         } else {
             yaml = new YamlConfiguration();
         }
-        Chicken.wing(main);
     }
 
     /**